add centrifugo plugin
