+++ .github/workflows/linux_on_push.yml
@@ -1074,134 +1074,6 @@ jobs:
           name: coverage
           path: ./coverage-ci/tcp.out
 
-  broadcast_test:
-    name: broadcast plugin (Go ${{ matrix.go }}, PHP ${{ matrix.php }}, OS ${{matrix.os}})
-    runs-on: ${{ matrix.os }}
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        php: [ "8.1" ]
-        go: [ "1.19" ]
-        os: [ "ubuntu-latest" ]
-    steps:
-      - name: Set up Go ${{ matrix.go }}
-        uses: actions/setup-go@v3 # action page: <https://github.com/actions/setup-go>
-        with:
-          go-version: ${{ matrix.go }}
-
-      - name: Set up PHP ${{ matrix.php }}
-        uses: shivammathur/setup-php@v2 # action page: <https://github.com/shivammathur/setup-php>
-        with:
-          php-version: ${{ matrix.php }}
-          extensions: sockets
-
-      - name: Check out code
-        uses: actions/checkout@v3
-
-      - name: Get Composer Cache Directory
-        id: composer-cache
-        run: echo "::set-output name=dir::$(composer config cache-files-dir)"
-
-      - name: Init Composer Cache # Docs: <https://git.io/JfAKn#php---composer>
-        uses: actions/cache@v3
-        with:
-          path: ${{ steps.composer-cache.outputs.dir }}
-          key: ${{ runner.os }}-composer-${{ matrix.php }}-${{ hashFiles('**/composer.json') }}
-          restore-keys: ${{ runner.os }}-composer-
-
-      - name: Install Composer dependencies
-        run: cd php_test_files && composer update --prefer-dist --no-progress --ansi
-
-      - name: Init Go modules Cache # Docs: <https://git.io/JfAKn#go---modules>
-        uses: actions/cache@v3
-        with:
-          path: ~/go/pkg/mod
-          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
-          restore-keys: ${{ runner.os }}-go-
-
-      - name: Install Go dependencies
-        run: go mod download
-
-      - name: Run golang tests with coverage
-        env:
-          RR_TEST_ENV: ${{ secrets.RR_TEST_ENV }}
-          RR_SQS_TEST_ENDPOINT: ${{ secrets.RR_SQS_TEST_ENDPOINT }}
-          RR_SQS_TEST_REGION: ${{ secrets.RR_SQS_TEST_REGION }}
-          RR_SQS_TEST_KEY: ${{ secrets.RR_SQS_TEST_KEY }}
-          RR_SQS_TEST_SECRET: ${{ secrets.RR_SQS_TEST_SECRET }}
-
-        run: |
-          docker-compose -f env/docker-compose-kv.yaml up -d --remove-orphans
-          mkdir ./coverage-ci
-          go test -timeout 20m -v -race -cover -tags=debug -failfast -coverpkg=all -coverprofile=./coverage-ci/broadcast.out -covermode=atomic ./plugins/broadcast
-
-      - name: Archive code coverage results
-        uses: actions/upload-artifact@v3
-        with:
-          name: coverage
-          path: ./coverage-ci/broadcast.out
-
-  websockets_test:
-    name: websockets plugin (Go ${{ matrix.go }}, PHP ${{ matrix.php }}, OS ${{matrix.os}})
-    runs-on: ${{ matrix.os }}
-    timeout-minutes: 60
-    strategy:
-      matrix:
-        php: [ "8.1" ]
-        go: [ "1.19" ]
-        os: [ "ubuntu-latest" ]
-    steps:
-      - name: Set up Go ${{ matrix.go }}
-        uses: actions/setup-go@v3 # action page: <https://github.com/actions/setup-go>
-        with:
-          go-version: ${{ matrix.go }}
-
-      - name: Set up PHP ${{ matrix.php }}
-        uses: shivammathur/setup-php@v2 # action page: <https://github.com/shivammathur/setup-php>
-        with:
-          php-version: ${{ matrix.php }}
-          extensions: sockets
-
-      - name: Check out code
-        uses: actions/checkout@v3
-
-      - name: Get Composer Cache Directory
-        id: composer-cache
-        run: echo "::set-output name=dir::$(composer config cache-files-dir)"
-
-      - name: Init Composer Cache # Docs: <https://git.io/JfAKn#php---composer>
-        uses: actions/cache@v3
-        with:
-          path: ${{ steps.composer-cache.outputs.dir }}
-          key: ${{ runner.os }}-composer-${{ matrix.php }}-${{ hashFiles('**/composer.json') }}
-          restore-keys: ${{ runner.os }}-composer-
-
-      - name: Install Composer dependencies
-        run: cd php_test_files && composer update --prefer-dist --no-progress --ansi
-
-      - name: Init Go modules Cache # Docs: <https://git.io/JfAKn#go---modules>
-        uses: actions/cache@v3
-        with:
-          path: ~/go/pkg/mod
-          key: ${{ runner.os }}-go-${{ hashFiles('**/go.sum') }}
-          restore-keys: ${{ runner.os }}-go-
-
-      - name: Install Go dependencies
-        run: go mod download
-
-      - name: Run golang tests with coverage
-        run: |
-          docker-compose -f env/docker-compose-kv.yaml up -d --remove-orphans
-          sleep 30
-          mkdir ./coverage-ci
-          go test -v -race -cover -tags=debug -failfast -coverpkg=all -coverprofile=./coverage-ci/websockets.out -covermode=atomic ./plugins/websockets
-
-      - name: Archive code coverage results
-        uses: actions/upload-artifact@v3
-        with:
-          name: coverage
-          path: ./coverage-ci/websockets.out
-
   http_test:
     name: http plugin (Go ${{ matrix.go }}, PHP ${{ matrix.php }}, OS ${{matrix.os}})
     runs-on: ${{ matrix.os }}
@@ -2074,8 +1946,6 @@ jobs:
       - kv_test
       - tcp_test
       - reload_test
-      - broadcast_test
-      - websockets_test
       - http_test
       - grpc_test
       - informer_test
