--- plugins/broadcast/configs/.rr-broadcast-same-section.yaml
@@ -1,40 +1,0 @@
-version: '2.7'
-
-rpc:
-    listen: tcp://127.0.0.1:6002
-
-server:
-    command: "php ../../php_test_files/psr-worker-bench.php"
-    relay: "pipes"
-    relay_timeout: "20s"
-
-http:
-    address: 127.0.0.1:28976
-    max_request_size: 1024
-    middleware: ["websockets"]
-    pool:
-        num_workers: 2
-        max_jobs: 0
-        allocate_timeout: 60s
-        destroy_timeout: 60s
-
-broadcast:
-    test:
-        driver: redis
-        config:
-            addrs:
-                - "127.0.0.1:6379"
-    test2:
-        driver: redis
-        config:
-            addrs:
-                - "127.0.0.1:6378"
-    test3:
-        driver: memory
-        config: {}
-    test4:
-        driver: memory
-        config: {}
-logs:
-    mode: development
-    level: info
