--- plugins/websockets/configs/.rr-websockets-deny.yaml
@@ -1,50 +1,0 @@
-version: '2.7'
-
-rpc:
-    listen: tcp://127.0.0.1:6001
-
-server:
-    command: "php ../../php_test_files/worker-deny.php"
-    user: ""
-    group: ""
-    relay: "pipes"
-    relay_timeout: "20s"
-
-http:
-    address: 127.0.0.1:15587
-    max_request_size: 1024
-    middleware: ["websockets"]
-    trusted_subnets:
-        [
-            "10.0.0.0/8",
-            "127.0.0.0/8",
-            "172.16.0.0/12",
-            "192.168.0.0/16",
-            "::1/128",
-            "fc00::/7",
-            "fe80::/10",
-        ]
-    pool:
-        num_workers: 2
-        max_jobs: 0
-        allocate_timeout: 60s
-        destroy_timeout: 60s
-
-broadcast:
-    test:
-        driver: memory
-        config: {}
-
-websockets:
-    broker: test
-    allowed_origin: "*"
-    path: "/ws"
-
-logs:
-    mode: development
-    level: error
-
-endure:
-    grace_period: 120s
-    print_graph: false
-    log_level: error
