--- plugins/websockets/websocket_plugin_test.go
@@ -1,853 +1,0 @@
-package websockets
-
-import (
-	"context"
-	"net"
-	"net/rpc"
-	"net/url"
-	"os"
-	"os/signal"
-	"sync"
-	"syscall"
-	"testing"
-	"time"
-
-	"github.com/gobwas/ws"
-	"github.com/gobwas/ws/wsutil"
-	"github.com/roadrunner-server/broadcast/v3"
-	"github.com/roadrunner-server/config/v3"
-	"github.com/roadrunner-server/logger/v3"
-	"github.com/roadrunner-server/memory/v3"
-	proxy "github.com/roadrunner-server/proxy_ip_parser/v3"
-	"github.com/roadrunner-server/redis/v3"
-	"github.com/roadrunner-server/server/v3"
-	"github.com/roadrunner-server/websockets/v3"
-	"github.com/stretchr/testify/require"
-
-	"github.com/goccy/go-json"
-	endure "github.com/roadrunner-server/endure/pkg/container"
-	goridgeRpc "github.com/roadrunner-server/goridge/v3/pkg/rpc"
-	httpPlugin "github.com/roadrunner-server/http/v3"
-	rpcPlugin "github.com/roadrunner-server/rpc/v3"
-	"github.com/roadrunner-server/sdk/v3/utils"
-	"github.com/stretchr/testify/assert"
-	websocketsProto "go.buf.build/protocolbuffers/go/roadrunner-server/api/proto/websockets/v1"
-)
-
-func TestWebsocketsInit(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-init.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&proxy.Plugin{},
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&memory.Plugin{},
-		&broadcast.Plugin{},
-	)
-
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Minute)
-	time.Sleep(time.Second * 1)
-	t.Run("TestWSInit", wsInit)
-	t.Run("RPCWsMemoryPubAsync", RPCWsPubAsync("11111"))
-	t.Run("RPCWsMemory", RPCWsPub("11111"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func TestWSRedis(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-redis.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsRedisPubAsync", RPCWsPubAsync("13235"))
-	t.Run("RPCWsRedisPub", RPCWsPub("13235"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func TestWSRedisNoSection(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-broker-no-section.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	_, err = cont.Serve()
-	assert.Error(t, err)
-}
-
-func TestWSDeny(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-deny.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&memory.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsMemoryDeny", RPCWsDeny("15587"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func TestWSDeny2(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-deny2.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&redis.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsRedisDeny", RPCWsDeny("15588"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func TestWSStop(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-stop.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&memory.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsStop", RPCWsMemoryStop("11114"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func RPCWsMemoryStop(port string) func(t *testing.T) {
-	return func(t *testing.T) {
-		connURL := url.URL{Scheme: "ws", Host: "127.0.0.1:" + port, Path: "/ws"}
-
-		_, _, _, err := ws.Dial(context.Background(), connURL.String())
-		require.Error(t, err)
-		if nErr, ok := err.(ws.StatusError); ok {
-			require.Equal(t, "unexpected HTTP response status: 401", nErr.Error())
-		}
-	}
-}
-
-func TestWSAllow(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-allow.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&memory.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsMemoryAllow", RPCWsPub("41278"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func TestWSAllow2(t *testing.T) {
-	cont, err := endure.NewContainer(nil, endure.SetLogLevel(endure.ErrorLevel))
-	assert.NoError(t, err)
-
-	cfg := &config.Plugin{
-		Version: "2.9.0",
-		Path:    "configs/.rr-websockets-allow2.yaml",
-		Prefix:  "rr",
-	}
-
-	err = cont.RegisterAll(
-		cfg,
-		&rpcPlugin.Plugin{},
-		&logger.Plugin{},
-		&server.Plugin{},
-		&redis.Plugin{},
-		&websockets.Plugin{},
-		&httpPlugin.Plugin{},
-		&memory.Plugin{},
-		&broadcast.Plugin{},
-	)
-	assert.NoError(t, err)
-
-	err = cont.Init()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	ch, err := cont.Serve()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-	wg := &sync.WaitGroup{}
-	wg.Add(1)
-
-	stopCh := make(chan struct{}, 1)
-
-	go func() {
-		defer wg.Done()
-		for {
-			select {
-			case e := <-ch:
-				assert.Fail(t, "error", e.Error.Error())
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-			case <-sig:
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			case <-stopCh:
-				// timeout
-				err = cont.Stop()
-				if err != nil {
-					assert.FailNow(t, "error", err.Error())
-				}
-				return
-			}
-		}
-	}()
-
-	time.Sleep(time.Second * 1)
-	t.Run("RPCWsMemoryAllow", RPCWsPub("41270"))
-
-	stopCh <- struct{}{}
-
-	wg.Wait()
-}
-
-func wsInit(t *testing.T) {
-	connURL := url.URL{Scheme: "ws", Host: "127.0.0.1:11111", Path: "/ws"}
-	dialer := ws.Dialer{
-		Header: ws.HandshakeHeaderHTTP{"Origin": []string{"127.0.0.1"}},
-	}
-
-	conn, _, _, err := dialer.Dial(context.Background(), connURL.String())
-	require.NoError(t, err)
-
-	d, err := json.Marshal(messageWS("join", []byte("hello websockets"), "foo", "foo2"))
-	require.NoError(t, err)
-
-	err = wsutil.WriteClientText(conn, d)
-	assert.NoError(t, err)
-
-	msg, err := wsutil.ReadServerText(conn)
-	require.NoError(t, err)
-	_ = msg
-	retMsg := utils.AsString(msg)
-
-	// subscription done
-	assert.Equal(t, `{"topic":"@join","payload":["foo","foo2"]}`, retMsg)
-
-	err = conn.Close()
-	require.NoError(t, err)
-}
-
-func RPCWsPubAsync(port string) func(t *testing.T) {
-	return func(t *testing.T) {
-		connURL := url.URL{Scheme: "ws", Host: "127.0.0.1:" + port, Path: "/ws"}
-		conn, _, _, err := ws.Dial(context.Background(), connURL.String())
-		require.NoError(t, err)
-
-		go func() {
-			messagesToVerify := make([]string, 0, 10)
-			messagesToVerify = append(messagesToVerify, `{"topic":"@join","payload":["foo","foo2"]}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"foo","payload":"hello, PHP"}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"@leave","payload":["foo"]}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"foo2","payload":"hello, PHP2"}`)
-			i := 0
-			for {
-				msg, err2 := wsutil.ReadServerText(conn)
-				retMsg := utils.AsString(msg)
-				assert.NoError(t, err2)
-				assert.Equal(t, messagesToVerify[i], retMsg)
-				i++
-				if i == 3 {
-					return
-				}
-			}
-		}()
-
-		time.Sleep(time.Second)
-
-		d, err := json.Marshal(messageWS("join", []byte("hello websockets"), "foo", "foo2"))
-		require.NoError(t, err)
-
-		err = wsutil.WriteClientText(conn, d)
-		require.NoError(t, err)
-
-		time.Sleep(time.Second)
-
-		publishAsync(t, "foo")
-
-		time.Sleep(time.Second)
-
-		// //// LEAVE foo /////////
-		d, err = json.Marshal(messageWS("leave", []byte("hello websockets"), "foo"))
-		if err != nil {
-			panic(err)
-		}
-
-		err = wsutil.WriteClientText(conn, d)
-		require.NoError(t, err)
-
-		time.Sleep(time.Second)
-
-		// TRY TO PUBLISH TO UNSUBSCRIBED TOPIC
-		publishAsync(t, "foo")
-
-		err = conn.Close()
-		assert.NoError(t, err)
-	}
-}
-
-func RPCWsPub(port string) func(t *testing.T) {
-	return func(t *testing.T) {
-		connURL := url.URL{Scheme: "ws", Host: "127.0.0.1:" + port, Path: "/ws"}
-
-		conn, _, _, err := ws.Dial(context.Background(), connURL.String())
-		require.NoError(t, err)
-
-		go func() {
-			messagesToVerify := make([]string, 0, 10)
-			messagesToVerify = append(messagesToVerify, `{"topic":"@join","payload":["foo","foo2"]}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"foo","payload":"hello, PHP"}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"@leave","payload":["foo"]}`)
-			messagesToVerify = append(messagesToVerify, `{"topic":"foo2","payload":"hello, PHP2"}`)
-			i := 0
-			for {
-				msg, err2 := wsutil.ReadServerText(conn)
-				retMsg := utils.AsString(msg)
-				assert.NoError(t, err2)
-				assert.Equal(t, messagesToVerify[i], retMsg)
-				i++
-				if i == 3 {
-					return
-				}
-			}
-		}()
-
-		time.Sleep(time.Second)
-
-		d, err := json.Marshal(messageWS("join", []byte("hello websockets"), "foo", "foo2"))
-		if err != nil {
-			panic(err)
-		}
-
-		err = wsutil.WriteClientText(conn, d)
-		assert.NoError(t, err)
-
-		time.Sleep(time.Second)
-
-		publish("", "foo")
-
-		time.Sleep(time.Second)
-
-		// //// LEAVE foo /////////
-		d, err = json.Marshal(messageWS("leave", []byte("hello websockets"), "foo"))
-		if err != nil {
-			panic(err)
-		}
-
-		err = wsutil.WriteClientText(conn, d)
-		assert.NoError(t, err)
-
-		time.Sleep(time.Second)
-
-		// TRY TO PUBLISH TO UNSUBSCRIBED TOPIC
-		publish("", "foo")
-
-		err = conn.Close()
-		assert.NoError(t, err)
-	}
-}
-
-func RPCWsDeny(port string) func(t *testing.T) {
-	return func(t *testing.T) {
-		connURL := url.URL{Scheme: "ws", Host: "127.0.0.1:" + port, Path: "/ws"}
-		conn, _, _, err := ws.Dial(context.Background(), connURL.String())
-		require.NoError(t, err)
-
-		d, err := json.Marshal(messageWS("join", []byte("hello websockets"), "foo", "foo2"))
-		if err != nil {
-			panic(err)
-		}
-
-		err = wsutil.WriteClientText(conn, d)
-		assert.NoError(t, err)
-
-		msg, err := wsutil.ReadServerText(conn)
-		require.NoError(t, err)
-		retMsg := utils.AsString(msg)
-
-		// subscription done
-		assert.Equal(t, `{"topic":"#join","payload":["foo","foo2"]}`, retMsg)
-
-		// //// LEAVE foo, foo2 /////////
-		d, err = json.Marshal(messageWS("leave", []byte("hello websockets"), "foo"))
-		if err != nil {
-			panic(err)
-		}
-
-		err = wsutil.WriteClientText(conn, d)
-		assert.NoError(t, err)
-
-		msg, err = wsutil.ReadServerText(conn)
-		require.NoError(t, err)
-		retMsg = utils.AsString(msg)
-
-		// subscription done
-		assert.Equal(t, `{"topic":"@leave","payload":["foo"]}`, retMsg)
-
-		err = conn.Close()
-		assert.NoError(t, err)
-	}
-}
-
-// ---------------------------------------------------------------------------------------------------
-
-func publish(topics ...string) {
-	conn, err := net.Dial("tcp", "127.0.0.1:6001")
-	if err != nil {
-		panic(err)
-	}
-
-	client := rpc.NewClientWithCodec(goridgeRpc.NewClientCodec(conn))
-
-	ret := &websocketsProto.Response{}
-	err = client.Call("broadcast.Publish", makeMessage([]byte("hello, PHP"), topics...), ret)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func publishAsync(t *testing.T, topics ...string) {
-	conn, err := net.Dial("tcp", "127.0.0.1:6001")
-	if err != nil {
-		panic(err)
-	}
-
-	client := rpc.NewClientWithCodec(goridgeRpc.NewClientCodec(conn))
-
-	ret := &websocketsProto.Response{}
-	err = client.Call("broadcast.PublishAsync", makeMessage([]byte("hello, PHP"), topics...), ret)
-	assert.NoError(t, err)
-	assert.True(t, ret.Ok)
-}
-
-func messageWS(command string, payload []byte, topics ...string) *websocketsProto.Message {
-	return &websocketsProto.Message{
-		Topics:  topics,
-		Command: command,
-		Payload: payload,
-	}
-}
-
-func makeMessage(payload []byte, topics ...string) *websocketsProto.Request {
-	m := &websocketsProto.Request{
-		Messages: []*websocketsProto.Message{
-			{
-				Topics:  topics,
-				Payload: payload,
-			},
-		},
-	}
-
-	return m
-}
