+++ test-certs/rootCA.pem
@@ -1,27 +1,29 @@
 -----BEGIN CERTIFICATE-----
-MIIEiTCCAvGgAwIBAgIQbIUZlGmNOzAkldDFs+gy8TANBgkqhkiG9w0BAQsFADBd
-MR4wHAYDVQQKExVta2NlcnQgZGV2ZWxvcG1lbnQgQ0ExGTAXBgNVBAsMEHZhbGVy
-eUBydXN0YXRpYW4xIDAeBgNVBAMMF21rY2VydCB2YWxlcnlAcnVzdGF0aWFuMB4X
-DTIyMDkwODExMjYzOFoXDTMyMDkwODExMjYzOFowXTEeMBwGA1UEChMVbWtjZXJ0
-IGRldmVsb3BtZW50IENBMRkwFwYDVQQLDBB2YWxlcnlAcnVzdGF0aWFuMSAwHgYD
-VQQDDBdta2NlcnQgdmFsZXJ5QHJ1c3RhdGlhbjCCAaIwDQYJKoZIhvcNAQEBBQAD
-ggGPADCCAYoCggGBALc6r8nRbwcnSOHJxxEk0tQM1BO7Qj+zcbQ1u4sPVj3/DOaT
-3cGIgZXbvteijyugkyT8ya2zynwU4OFeJ7PSFk7ZoyXTKOKbPdAs8eAP7rYMPynJ
-Oq6v+HwJ0JVhpiZtyKF3QVlJ3hVO2bHuPD83JiFl0D6h3GXKeeDaze8dDgToogIT
-kOekIgzbFEUyyBNjj07XKVCWOLwDHph5/eKCVrP80oqj4fF0MeZDYpjkWBnaEYvq
-kmoMQcLl6z6nL5vG8xrs/LIbzUDWAvBkMQKcOf0C5jAd7xqAFOLd4VhGB3cnbr9u
-/uBvX/fmAVuJVGGRDx5ntF1NeVMjXXOPsj4vwDq7REjRMnkVodqjeP9VKz+PP4dU
-gYh98vjbiW4O6VbUaVn29kjBU7zYei4+YSfE8A9IliERKihABHSGOjeJffpkGbxu
-xSAhcCG/X4Faxh0C7oJ7ijbSwGGJKIM8Voo9QCkVV/EQmhd/ResixzD7HfN4RJ5r
-RC3TZNrYTqERYJ+ONwIDAQABo0UwQzAOBgNVHQ8BAf8EBAMCAgQwEgYDVR0TAQH/
-BAgwBgEB/wIBADAdBgNVHQ4EFgQUayfGipyxW5LcXbETHFHeJu4zB6AwDQYJKoZI
-hvcNAQELBQADggGBACb/ASD24r9XZEudY+ZfEhaQb/kfyfxw0yNUlFrnRsht2yk6
-VZ5SpwOdu/R/oknDn5eKmZme1jEzeQgxbJiDUaiY+xn82iwc9vBgVUc4/eQoAMMQ
-Ht5GrFFDcQjEJlngqg3lwqrnjIbL2EThvGYLJjDdqoADgI+bk6p99VGEJTYc4Kcn
-HrNzPLfbqAsVhh4FH5Z6xx2jdwO2LEFg5KOD22SCXzB2khqbD+O8tNXR6sMIUL/Y
-6gpAmzYLIaOFgklBM3+IkO9WHCeSVLNgHU6ITXirNHET8QTXsJZRTXCJ4f630vVg
-6ZsucGMiFOPvCN6S2Bi4S1xjcp2qqy/s8Ljzfjs6oFCdj9HAVWrCBohncFkAMMud
-c5AunzEWHwEsIG970vnUFFO9+Yw+dTcVKB/TvORzjm8XBPfLtSx278l1I4oDuSIC
-5gKvfVfZbmSyn1WCK3rj9ArB4UjGPGarDijd0amMjY4xbP2H28YNlva9y4j+ZqRn
-bhnYD6W9DKowR+q2Xw==
+MIIE4zCCA0ugAwIBAgIQDwrPEDA905XY+cOcyV/cFDANBgkqhkiG9w0BAQsFADCB
+iTEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMS8wLQYDVQQLDCZ2YWxl
+cnlAcnVzdGF0aWFuIChWYWxlcnkgUGlhc2hjaHluc2tpKTE2MDQGA1UEAwwtbWtj
+ZXJ0IHZhbGVyeUBydXN0YXRpYW4gKFZhbGVyeSBQaWFzaGNoeW5za2kpMB4XDTIy
+MTAyMjEzNDkyOFoXDTMyMTAyMjEzNDkyOFowgYkxHjAcBgNVBAoTFW1rY2VydCBk
+ZXZlbG9wbWVudCBDQTEvMC0GA1UECwwmdmFsZXJ5QHJ1c3RhdGlhbiAoVmFsZXJ5
+IFBpYXNoY2h5bnNraSkxNjA0BgNVBAMMLW1rY2VydCB2YWxlcnlAcnVzdGF0aWFu
+IChWYWxlcnkgUGlhc2hjaHluc2tpKTCCAaIwDQYJKoZIhvcNAQEBBQADggGPADCC
+AYoCggGBALEcm9QVAb/cQKywLgHmEv9xBIGMWchYcxgunxK4mijTi/wQ1WoCsJae
+oQF3eawbYgyFxDrHxTW9NAGOA902v4SA8emM8ixqWK9NiwTGzQ8XdvICt3kjCs1X
+vPo8k//12PgUbc5kd5R0YSr2iHo8t0vV1+tWcNFHTRSkh87M/cLD6L9dBVyVSu2p
+RqmBKl3y6qi3HBjApbu9T22y9icS7Y8A6gyQbhpfE58ABDau7dpmWNg/ZUja7CSQ
+EZH3JxMN9daJAQgImZcOSoJuF3NMcehTyqeQVnlUFrzP0r0OXBesG2jF7E1HRwit
+OAayfDjxOAx+qeJVsw1Ws/8wmNdkhe0e/cAZl1QY5kjFptghlnhn9jNyUNLc0vwz
+rXkP7Y8mzeTig/BBTLCq2JuP+MGq1xzqX4tOgdLNo3xx384jCNxdhR7QB/5b5NQ9
+h8chAhowxjjSqaC0MzdM/wDcCSSNDWBgOV7KoXCFmMbs2svikmKaeePnFGBO+Rqg
+V/51o0lo0wIDAQABo0UwQzAOBgNVHQ8BAf8EBAMCAgQwEgYDVR0TAQH/BAgwBgEB
+/wIBADAdBgNVHQ4EFgQU8kyr8QwRRDFcLts90HDnaahuPQwwDQYJKoZIhvcNAQEL
+BQADggGBAEg+WLNqTl9OH4CiZtRVPNK+XbfFl+Uhqtn/+YMdX2ipH0p29bESF956
+qFKG3RlLlzW0JP2C31R/pWpQ0nc5pvBBMkt6o92h0Jj4GAZCn08HA2wbdeOD3b8N
+LnvtqeCvV7k1sX2VkTyWgjSmcHzNXy2cg7djHCeF46mXt/7a4pAjyTan9chOFiLD
+XJaqUFjv2xRRO6hOjysf7jOwqTWw4ut1Fy1z5BEmE7HSkz81S/s7bR9sOwWgVS6U
+WyryejvYsnReMYd6HBLgPaAQZZLI5PBxEbrAFLT3KPhU22s4A9mYX63Ik/Nue74S
+8NQhLOJHURLUEtfmdQDYvSA1gQ5T2WP7IPA/zDsjscIm+Z63Ucf241ipdLs64O+7
+f+ReOc7sfv3wl9eVri0RJh+JHgjsnQchD5kPHpkiaiGtjvKfzhg9w1DyHtHB2mPP
+hj10KZo2V1kfjxJ32G7SIeDBxV2uLwotbztJrO2uKdP+Tt56F95JjJ8TFmAO7heJ
+0NvvCcTtow==
 -----END CERTIFICATE-----
